Initial component loading implementation
- Migrated archive plugin into component plugin
