+++ flexget/components/archive/__init__.py
@@ -1,0 +1,3 @@
+# Import plugins when component is imported
+from . import cli
+from . import archive
